add functionality in search bar with tables

+++ Module-2/Ex-TedTalks/Ex-TedTalks/controller/ViewController.swift
@@ -11,15 +11,15 @@ import WebKit
 class ViewController: UIViewController {
     
     @IBOutlet weak var searchTable: UITableView!
-   
-   // @IBOutlet weak var TalksTableView: UITableView!
+    
+    @IBOutlet weak var TalksTableView: UITableView!
     @IBOutlet weak var searchBar: UISearchBar!
     var talks: [Talk] = []
     var filteredData: [Talk] = []
     var filteredTags: [String] = []
     override func viewDidLoad() {
         super.viewDidLoad()
-       // TalksTableView.separatorStyle = .none
+        TalksTableView.separatorStyle = .none
         TalkManager().retrieve(){
             [weak self] result in
             switch result{
@@ -33,26 +33,39 @@ class ViewController: UIViewController {
         }
         filteredData = talks
         filteredTags = filteredData.flatMap({ $0.tags})
-     
+        
     }
     
 }
 extension ViewController: UITableViewDataSource{
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        filteredData.count
+        guard tableView.tag == 1 else{
+            return filteredData.count
+        }
+        return filteredTags.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard tableView.tag == 1 else{
+            let cell = TalksTableView.dequeueReusableCell(withIdentifier: "TalkCell",for: indexPath) as!TalkCell
+            let videoUrl = "https://embed.ted.com/talks/majora_carter_s_tale_of_urban_renewal"
+            cell.video.load(URLRequest(url: URL(string: videoUrl)!))
+            cell.titleLabel.text = "Ken Robinson: Do schools kill creativity?"
+            cell.descriptionLabel.text = "With the same humor and humanity he exuded in 'An Inconvenient Truth,' Al Gore spells out 15 ways that individuals can address climate change immediately, from buying a hybrid to inventing a new, hotter brand name for global warming."
+           // cell.textLabel? = filteredData[indexPath.row].
+            return cell
+        }
         let cell = searchTable.dequeueReusableCell(withIdentifier: "searchCell",for: indexPath)
         cell.textLabel?.text = filteredTags[indexPath.row]
-        
-       /* let cell = TalksTableView.dequeueReusableCell(withIdentifier: "TalkCell",for: indexPath) as! TalkCell
-        let videoUrl = "https://embed.ted.com/talks/majora_carter_s_tale_of_urban_renewal"
-        cell.video.load(URLRequest(url: URL(string: videoUrl)!))
-        cell.titleLabel.text = "Ken Robinson: Do schools kill creativity?"
-        cell.descriptionLabel.text = "With the same humor and humanity he exuded in 'An Inconvenient Truth,' Al Gore spells out 15 ways that individuals can address climate change immediately, from buying a hybrid to inventing a new, hotter brand name for global warming."
-       */
         return cell
+    }
+    func tableView(_ tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
+       if tableView.tag == 1 {
+           
+            TalksTableView.reloadData()
+        }
+        
+        
     }
 }
 
@@ -69,7 +82,10 @@ extension ViewController: UISearchBarDelegate{
             for tag in word.tags{
                 if tag.uppercased().contains(searchText.uppercased()){
                     filteredData.append(word)
-                    filteredTags.append(tag)
+                    if !filteredTags.contains(tag){
+                        filteredTags.append(tag)
+                    }
+                    
                 }
             }
